+++ Models/IProduct.cs
@@ -2,9 +2,9 @@
     public interface IProduct {
         int Id { get; }
         string Sku { get; set; }
-        double Price { get; set; }
+        decimal Price { get; set; }
         bool IsDigital { get; set; }
-        double? ShippingCost { get; set; }
+        decimal? ShippingCost { get; set; }
         double Weight { get; set; }
     }
 }
