+++ Models/StateOrCountryTaxPartRecord.cs
@@ -8,7 +8,7 @@ namespace Nwazet.Commerce.Models {
 
         public virtual string State { get; set; }
         public virtual string Country { get; set; }
-        public virtual double Rate { get; set; }
+        public virtual decimal Rate { get; set; }
         public virtual int Priority { get; set; }
     }
 }
